allow override of SchemaDecoder
When an execute() call is made, there will likely be cases where the
caller knows better than the driver how best to deserialize data- I
guess this is the point to the ASSUME command in cassandra-cli and which
has been requested for cqlsh.

The simplest way to allow caller-controlled deserialization, it seems to
me, is to allow passing in a SchemaDecoder-like class as an optional
parameter to execute(). The custom decoder will be used until a new
query is made.

This adds that.