avoid relative in-package imports
they'll work, but in some versions of Python should produce a warning
(see PEP-328).  the marshal module makes this worse, since there is an
official Python module by the same name which will get imported instead
after the import rules change. if we can require Python 2.5 with this
library, we could just use the explicit relative-import syntax, but it
looks like everything else here is compatible with Python 2.4, so we
should probably maintain that as long as feasible.

Changed all relative imports within the cql package to absolute imports,
resolved using the 'cql' name.