fix error message on unknown CqlResultType
It's not at all clear how it ever happened, but a user had the
cursor.process_execution_results() method receive None as the
response.type one time (which generated "TypeError: cannot concatenate
'str' and 'NoneType' objects" instead of "unknown result type None",
because of the invalid expectation of stringliness herein fixed,
which wouldn't have been very much more helpful, but at least this is a
little bit more correct). User also saw a KsDef returned with None in
the 'name' attribute, suggesting something outside of python-cql being
badly broken.

so, not much of a problem, but woohoo.