Expose column name storage types
Since some clients need to know how a column name was deserialized,
and/or what the raw bytes were. This makes a change to the SchemaDecoder
interface, but unless clients are replacing SchemaDecoder wholesale,
it should be backwards-compatible.

As of now, the "storage type" so exposed will pretty much always be the
CF's comparator, except for columns corresponding to a row key, but
there isn't really any other way for a CQL client to inspect a
CF's comparator, and sometimes identifying the CF to be queried would
require parsing CQL.